+++ lib/model/wellness/ToDo.dart
@@ -23,7 +23,8 @@ import 'package:rokwire_plugin/utils/utils.dart';
 class ToDoItem {
   static final String _dateTimeFormat = 'yyyy-MM-ddTHH:mm:ssZ';
 
-  final String? id;
+  //TBD: DD - make it final again when we have backend APIs
+  String? id;
   final String? name;
   final ToDoCategory? category;
   final DateTime? dueDateTimeUtc;
@@ -32,7 +33,7 @@ class ToDoItem {
   final List<String>? workDays;
   final ToDoItemLocation? location;
   final String? description;
-  final bool isCompleted;
+  bool isCompleted;
 
   ToDoItem(
       {this.id,