+++ lib/ui/wellness/todo/WellnessManageToDoCategoriesPanel.dart
@@ -55,6 +55,12 @@ class _WellnessManageToDoCategoriesPanelState extends State<WellnessManageToDoCa
   }
 
   @override
+  void dispose() {
+    NotificationService().unsubscribe(this);
+    super.dispose();
+  }
+
+  @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: HeaderBar(title: Localization().getStringEx('panel.wellness.categories.manage.title', 'Manage Categories')),
@@ -260,6 +266,14 @@ class _WellnessManageToDoCategoriesPanelState extends State<WellnessManageToDoCa
   }
 
   void _onTapDeleteCategory(ToDoCategory category) {
+    AppAlert.showConfirmationDialog(
+        buildContext: context,
+        message: Localization().getStringEx(
+            'panel.wellness.categories.manage.category.delete.confirmation.msg', 'Are sure that you want to delete this category?'),
+        positiveCallback: () => _deleteCategory(category));
+  }
+
+  void _deleteCategory(ToDoCategory category) {
     _setLoading(true);
     Wellness().deleteToDoCategoryCached(category.id!).then((success) {
       late String msg;